requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
redis>=5.0.1
psycopg2-binary>=2.9.10
jinja2>=3.1.2